Move Logging class to utils
This is to avoid module shadowing
